@@ -3,34 +3,31 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
-    "start": "nuxt start",
-    "generate": "nuxt generate"
+    "dev": "nuxi dev",
+    "build": "nuxi build",
+    "start": "nuxi preview",
+    "generate": "nuxi generate"
   },
   "dependencies": {
     "@inotom/vue-go-top": "^1.3.0",
-    "@nuxtjs/axios": "^5.13.6",
-    "@nuxtjs/sitemap": "^2.4.0",
+    "@nuxtjs/i18n": "^9.2.1",
+    "@nuxtjs/sitemap": "^7.2.6",
+    "@pinia/nuxt": "^0.10.1",
     "@splidejs/vue-splide": "^0.3.5",
     "animate.css": "^4.1.1",
     "core-js": "^3.25.3",
     "dropzone": "^6.0.0-beta.2",
     "fslightbox-vue": "^2.1.3",
     "gsap": "^3.11.4",
+    "h3": "^1.15.1",
     "mapbox-gl": "^3.5.2",
-    "nuxt": "^2.15.8",
-    "nuxt-i18n": "^6.28.1",
+    "nuxt": "^3.15.4",
     "nuxt-swiper": "^1.2.2",
-    "nuxt-vue-select": "^0.1.4",
-    "swiper": "^5.4.5",
-    "vue": "^2.7.10",
-    "vue-awesome-swiper": "^4.1.1",
-    "vue-burger-menu": "^2.0.5",
+    "swiper": "^11.2.6",
+    "vue": "^3.5.13",
     "vue-scroll-snap": "^0.1.11",
     "vue-server-renderer": "^2.7.10",
-    "vue-template-compiler": "^2.7.10",
-    "vue2-google-maps": "^0.10.7"
+    "vue-template-compiler": "^2.7.10"
   },
   "devDependencies": {
     "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
@@ -38,9 +35,18 @@
     "@babel/preset-env": "^7.24.8",
     "@nuxt/types": "~2.17.0",
     "sass": "^1.85.0",
-    "sass-loader": "^10.5.2"
+    "sass-loader": "^10.5.2",
+    "typescript": "^5.8.2"
   },
   "overrides": {
     "autoprefixer": "10.4.5"
+  },
+  "optionalDependencies": {
+    "@rollup/rollup-linux-arm64-gnu": "4.9.5",
+    "@rollup/rollup-linux-x64-gnu": "4.9.5",
+    "@rollup/rollup-linux-x64-musl": "4.9.5"
+  },
+  "engines": {
+    "node": ">=20.0.0 <21.0.0"
   }
 }
