@@ -1,67 +1,55 @@
 <template>
-    <div :data-theme="darkMode" v-bind:class="['wrap', { 'wrap_small': this.$store.state.openMenu }]">
-        <SecTheHeader />
-        <div class="wrap__r">
-            <Nuxt class="wrapP" />
-        </div>
-        <CompScrollTop />
-        <SecTheFooter />
-    </div>
+  <div>
+      <NuxtLoadingIndicator color="#000" />
+      <SecTheHeader />
+      <div class="wrap__r">
+        <slot />
+      </div>
+      <CompScrollTop />
+      <SecTheFooter />
+  </div>
 </template>
-<script>
-
-export default {
-    name: "LayDef",
-    mounted() {
-        const isMenuOpen = this.getInitialIsMenuOpen();
-        this.$store.dispatch("updateOpenMenu", isMenuOpen);
-    },
-    methods: {
-        getInitialIsMenuOpen() {
-            if (typeof localStorage !== 'undefined') {
-                const storedValue = localStorage.getItem('isMenuOpen');
-                return storedValue !== null ? storedValue === 'true' : false;
-            }
-
-            return false;
-        },
-    },
-}
-
-</script>
-
 
 <script setup>
-import { ref } from 'vue';
+useHead({
+  title: "Kerimov",
+  meta: [
+    { charset: "utf-8" },
+    { name: "viewport", content: "width=device-width, initial-scale=1, viewport-fit=cover" },
+    { name: "p:domain_verify", content: "035b3120729cebc8cf517e99dcc25e99" },
+    { hid: "description", name: "description", content: "" },
+    { name: "format-detection", content: "telephone=no" },
+    { "http-equiv": "Content-Security-Policy", content: "upgrade-insecure-requests" },
+    { name: "robots", content: "index, follow" },
+  ],
+  link: [
+    { rel: "icon", type: "image/x-icon", href: "/favicon.ico" },
+  ],
+  script: [
+    {
+      innerHTML: `(function(m,e,t,r,i,k,a){m[i]=m[i]||function(){(m[i].a=m[i].a||[]).push(arguments)};
+      m[i].l=1*new Date();
+      for (var j = 0; j < document.scripts.length; j++) {if (document.scripts[j].src === r) { return; }}
+      k=e.createElement(t),a=e.getElementsByTagName(t)[0],k.async=1,k.src=r,a.parentNode.insertBefore(k,a)})
+      (window, document, "script", "https://mc.yandex.ru/metrika/tag.js", "ym");
 
-const getInitialDarkMode = () => {
-    if (typeof localStorage !== 'undefined') {
-        const storedValue = localStorage.getItem('darkMode');
-        return storedValue !== null ? storedValue === 'true' : false;
-    }
-    return false;
-};
-
-const darkMode = ref(getInitialDarkMode());
-
-const saveDarkModePreference = (isDarkMode) => {
-    if (typeof localStorage !== 'undefined') {
-        localStorage.setItem('darkMode', isDarkMode.toString());
-    }
-};
-
-const toggleDarkMode = () => {
-    darkMode.value = !darkMode.value;
-    saveDarkModePreference(darkMode.value);
-};
-
-if (typeof localStorage !== 'undefined') {
-    window.addEventListener('beforeunload', () => {
-        saveDarkModePreference(darkMode.value);
-    });
-}
-
-
+      ym(97429270, "init", {
+           clickmap:true,
+           trackLinks:true,
+           accurateTrackBounce:true,
+           webvisor:true
+      });`,
+      type: "text/javascript",
+      charset: "utf-8"
+    },
+  ],
+  noscript: [
+    {
+      innerHTML: '<div><img src="https://mc.yandex.ru/watch/97429270" style="position:absolute; left:-9999px;" alt="" /></div>',
+      body: true,
+    },
+  ],
+});
 </script>
 
 <style lang="scss">
@@ -135,6 +123,9 @@ if (typeof localStorage !== 'undefined') {
     flex: auto;
 }
 
+.wrap__r {
+  margin-top: 48px;
+}
 
 // .wrap__r {
 //     position: fixed;
@@ -171,10 +162,15 @@ if (typeof localStorage !== 'undefined') {
   backdrop-filter: saturate(180%) blur(20px);
   color: var(--black);
   z-index: 100;
+
+  &__cont {
+    margin-top: 48px;
+  }
 }
 
 .menuLogo {
   margin-top: 45px;
+  margin-bottom: 20px;
 
   svg {
       width: 80%;
